Extract more code from causal_tree.go
