+++ src/causal_tree/causal_tree.go
@@ -1,84 +1,23 @@
-/*
-Package crdt provides primitives to operate on replicated data types.
-
-Replicated data types are structured such that they can be copied across multiple sites
-in a distributed environment, mutated independently at each site, and they still may be
-merged back without conflicts.
-
-This implementation is based on the Causal Tree structure proposed by Victor Grishchenko [1],
-following the excellent explanation by Archagon [2].
-
-[1]: GRISCHENKO, VICTOR. Causal trees: towards real-time read-write hypertext.
-[2]: http://archagon.net/blog/2018/03/24/data-laced-with-history/
-*/
 package causal_tree
 
 import (
 	"bytes"
-	"crypto/rand"
 	"encoding/json"
 	"errors"
 	"fmt"
-	"io"
 	"math"
 	"sort"
-	"strconv"
-	"strings"
 
 	atm "github.com/crdteam/causal-tree/src/atom"
+	"github.com/crdteam/causal-tree/src/utils/conversion"
+	"github.com/crdteam/causal-tree/src/utils/generate_uuid_elements"
 	"github.com/crdteam/causal-tree/src/utils/indexmap"
+	wft "github.com/crdteam/causal-tree/src/weft"
 	"github.com/google/uuid"
 )
 
-/*
-In a causal tree, each operation on the data structure is represented by an atom, which
-has a single other atom as its cause, thus creating a tree shape.
-When inserting a char, for example, its cause is the char to its left.
-
-  # BEGIN ASCII ART
-
-  T <- H <- I <- S <- _ <- I <- S <- _ <- N <- I <- C <- E
-                                     ^
-                                     '-- V <- E <- R <- Y <- _
-
-  # END ASCII ART
-  # ALT TEXT: Sequence of letters with arrows between them, representing atoms and their causes.
-              In the first line it reads "THIS_IS_NICE", and in the second the string "VERY_" points
-              to the space after "IS". This represents an insertion, thus the whole tree should be
-              read as "THIS_IS_VERY_NICE".
-
-Instead of using a pointer to reference the causing operation, references simply hold an atom
-ID containing the origin site and the (local) timestamp of creation.
-Atoms are then organized in an array to improve memory locality, at the expanse of having to
-search linearly for a given atom ID.
-
-By sorting the array such that atoms from the same site and time are mostly contiguous, this search
-operation is not terribly costly, and the array reads almost like the structure being represented.
-
-  # BEGIN ASCII ART
-
-   id cause                                  .--------------------------------------.
-   |  |                                      .--------.                             |
-   v  v                                      v        |                             |
-  .-----.-----.-----.-----.-----.-----.-----.-----.-----.-----.-----.-----.-----.-----.-----.-----.-----.
-  |01|__|02|01|03|02|04|03|05|04|06|05|07|06|08|07|13|08|14|13|15|14|16|15|17|16|09|08|10|09|11|10|12|11|
-  +-----+-----+-----+-----+-----+-----+-----+-----+-----+-----+-----+-----+-----+-----+-----+-----+-----+
-  |  T  |  H  |  I  |  S  |  _  |  I  |  S  |  _  |  V  |  E  |  R  |  Y  |  _  |  N  |  I  |  C  |  E  |
-  '-----'-----'-----'-----'-----'-----'-----'-----'-----'-----'-----'-----'-----'-----'-----'-----'-----'
-
-  # END ASCII ART
-  # ALT TEXT: Sequence of atoms in an array, representing the same string as before. Each atom is composed
-              of three elements: ID, cause, and content. IDs are given sequentially, from 01 to 17.
-              From 01 to 08, all atoms are sorted by ID, with contents that spell "THIS_IS_".
-              To the right of 08, we have the sequence of IDs from 13 to 17, with contents that spell
-              "VERY_". The cause of ID 13 is the atom 08.
-              Finally, to the right of ID 17 is the sequence of atoms from 09 to 12, with contents that
-              spell "NICE". The cause of ID 09 is also the atom 08.
-              The first atom has no cause, and all others have as cause the atom to its left.
-*/
-
 var (
-	uuidv1 = randomUUIDv1 // Stubbed for mocking in mocks_test.go
+	uuidv1 = generate_uuid_elements.RandomUUIDv1 // Stubbed for mocking in mocks_test.go
 )
 
 // +-----------------------+
@@ -102,19 +41,6 @@ type CausalTree struct {
 	SiteID uuid.UUID
 	// Timestamp is this tree's Lamport timestamp.
 	Timestamp uint32
-}
-
-// NewCausalTree creates an initialized empty replicated tree.
-func NewCausalTree() *CausalTree {
-	siteID := uuidv1()
-	return &CausalTree{
-		Weave:     nil,
-		Cursor:    atm.AtomID{},
-		Yarns:     [][]atm.Atom{nil},
-		Sitemap:   []uuid.UUID{siteID},
-		SiteID:    siteID,
-		Timestamp: 1, // Timestamp 0 is considered invalid.
-	}
 }
 
 // Returns the index where a site is (or should be) in the sitemap.
@@ -369,41 +295,6 @@ func (t *CausalTree) fixDeletedCursor() {
 // + Time travel |
 // +-------------+
 
-// Weft is a clock that stores the timestamp of each site of a CausalTree.
-//
-// In a distributed system it's not possible to observe the whole state at an absolute time,
-// but we can view the site's state at each site time.
-type Weft []uint32
-
-// Compare returns -1, +1 and 0 if this is weft is less than, greater than, or concurrent
-// to the other, respectively.
-//
-// It panics if wefts have different sizes.
-func (w Weft) Compare(other Weft) int {
-	if len(w) != len(other) {
-		panic(fmt.Sprintf("wefts have different sizes: %d (%v) != %d (%v)", len(w), w, len(other), other))
-	}
-	var hasLess, hasGreater bool
-	for i, t1 := range w {
-		t2 := other[i]
-		if t1 < t2 {
-			hasLess = true
-		} else if t1 > t2 {
-			hasGreater = true
-		}
-	}
-	if hasLess && hasGreater {
-		return 0
-	}
-	if hasLess {
-		return -1
-	}
-	if hasGreater {
-		return +1
-	}
-	return 0
-}
-
 // The same as weft, but using yarn's indices instead of timestamps.
 type indexWeft []int
 
@@ -417,7 +308,7 @@ func (ixs indexWeft) isInView(id atm.AtomID) bool {
 // in other sites.
 //
 // Time complexity: O(atoms)
-func (t *CausalTree) checkWeft(weft Weft) (indexWeft, error) {
+func (t *CausalTree) checkWeft(weft wft.Weft) (indexWeft, error) {
 	if len(t.Yarns) != len(weft) {
 		return nil, ErrWeftInvalidLength
 	}
@@ -449,8 +340,8 @@ func (t *CausalTree) checkWeft(weft Weft) (indexWeft, error) {
 }
 
 // Now returns the last known time at every site as a weft.
-func (t *CausalTree) Now() Weft {
-	weft := make(Weft, len(t.Yarns))
+func (t *CausalTree) Now() wft.Weft {
+	weft := make(wft.Weft, len(t.Yarns))
 	for i, yarn := range t.Yarns {
 		n := len(yarn)
 		if n == 0 {
@@ -464,7 +355,7 @@ func (t *CausalTree) Now() Weft {
 // ViewAt returns a view of the tree in the provided time in the past, represented with a weft.
 //
 // Time complexity: O(atoms+sites)
-func (t *CausalTree) ViewAt(weft Weft) (*CausalTree, error) {
+func (t *CausalTree) ViewAt(weft wft.Weft) (*CausalTree, error) {
 	limits, err := t.checkWeft(weft)
 	if err != nil {
 		return nil, err
@@ -657,224 +548,9 @@ func (t *CausalTree) SetCursor(i int) error {
 	return nil
 }
 
-// +--------------------------------------+
-// + Operations - Atom Priority constants |
-// +--------------------------------------+
-const (
-	insertCharPriority    = 0
-	insertStrPriority     = 30
-	deletePriority        = 100
-	insertCounterPriority = 30
-	insertAddPriority     = 30
-)
-
-// +--------------------------+
-// | Operations - Insert char |
-// +--------------------------+
-
-// InsertChar represents insertion of a char to the right of another atom.
-type InsertChar struct {
-	// Char inserted in tree.
-	Char rune
-}
-
-func (v InsertChar) AtomPriority() int { return insertCharPriority }
-func (v InsertChar) MarshalJSON() ([]byte, error) {
-	return json.Marshal(fmt.Sprintf("insert %c", v.Char))
-}
-func (v InsertChar) String() string { return string([]rune{v.Char}) }
-
-func (v InsertChar) ValidateChild(child atm.AtomValue) error {
-	switch child.(type) {
-	case InsertChar, Delete:
-		return nil
-	default:
-		return fmt.Errorf("invalid atom value after InsertChar: %T (%v)", child, child)
-	}
-}
-
-// InsertChar inserts a char after the cursor position and advances the cursor.
-func (t *CausalTree) InsertChar(ch rune) error {
-	atomID, err := t.addAtom(InsertChar{ch})
-	if err != nil {
-		return err
-	}
-	t.Cursor = atomID
-	return nil
-}
-
-// InsertCharAt inserts a char after the given (tree) position.
-func (t *CausalTree) InsertCharAt(ch rune, i int) error {
-	if err := t.SetCursor(i); err != nil {
-		return err
-	}
-	return t.InsertChar(ch)
-}
-
-// +---------------------+
-// | Operations - Delete |
-// +---------------------+
-
-// Delete represents deleting an element from the tree.
-type Delete struct{}
-
-func (v Delete) AtomPriority() int { return deletePriority }
-func (v Delete) MarshalJSON() ([]byte, error) {
-	return []byte(`"delete"`), nil
-}
-func (v Delete) String() string { return "⌫ " }
-
-func (v Delete) ValidateChild(child atm.AtomValue) error {
-	return fmt.Errorf("invalid atom value after Delete: %T (%v)", child, child)
-}
-
-// Delete deletes the char at the cursor position, and relocates the cursor to its cause.
-func (t *CausalTree) Delete() error {
-	if t.Cursor.Timestamp == 0 {
-		return ErrNoAtomToDelete
-	}
-	if _, err := t.addAtom(Delete{}); err != nil {
-		return err
-	}
-	t.fixDeletedCursor()
-	return nil
-}
-
-// DeleteAt deletes the char at the given (tree) position.
-func (t *CausalTree) DeleteAt(i int) error {
-	if err := t.SetCursor(i); err != nil {
-		return err
-	}
-	return t.Delete()
-}
-
-// +-----------------------------------+
-// | Operations - Insert str container |
-// +-----------------------------------+
-
-//Inserts a string container as a child of the root atom.
-type InsertStr struct{}
-
-func (v InsertStr) AtomPriority() int { return insertStrPriority }
-func (v InsertStr) MarshalJSON() ([]byte, error) {
-	return json.Marshal("insert str container")
-}
-
-func (v InsertStr) String() string { return "STR: " }
-
-func (v InsertStr) ValidateChild(child atm.AtomValue) error {
-	switch child.(type) {
-	case InsertChar, Delete:
-		return nil
-	default:
-		return fmt.Errorf("invalid atom value after InsertStr: %T (%v)", child, child)
-	}
-}
-
-// InsertStr inserts a Str container after the root and advances the cursor.
-func (t *CausalTree) InsertStr() error {
-	t.Cursor = atm.AtomID{}
-	atomID, err := t.addAtom(InsertStr{})
-	t.Cursor = atomID
-	return err
-}
-
-// +------------------------------+
-// | Operations - Insert Add atom |
-// +------------------------------+
-
-//Inserts an add atom as a child of the atom pointed by cursor.
-type InsertAdd struct {
-	//Value inserted into the counter container
-	Value int32
-}
-
-func (v InsertAdd) AtomPriority() int { return insertAddPriority }
-func (v InsertAdd) MarshalJSON() ([]byte, error) {
-	return json.Marshal(fmt.Sprintf("insert %d", v.Value))
-}
-
-func (v InsertAdd) String() string { return strconv.FormatInt(int64(v.Value), 10) }
-
-//InsertAdd atoms only accept child of type InsertAdd.
-func (v InsertAdd) ValidateChild(child atm.AtomValue) error {
-	switch child.(type) {
-	case InsertAdd:
-		return nil
-	default:
-		return fmt.Errorf("invalid atom value after InsertAdd: %T (%v)", child, child)
-	}
-}
-
-// InsertAdd inserts an InsertAdd atom after the cursor position and advances the cursor.
-func (t *CausalTree) InsertAdd(val int32) error {
-	atomID, err := t.addAtom(InsertAdd{val})
-	if err != nil {
-		return err
-	}
-	t.Cursor = atomID
-	return nil
-}
-
-// InsertAddAt inserts an InsertAdd atom after the given (tree) position.
-func (t *CausalTree) InsertAddAt(val int32, i int) error {
-	if err := t.SetCursor(i); err != nil {
-		return err
-	}
-	return t.InsertAdd(val)
-}
-
-// +---------------------------------------+
-// | Operations - Insert counter container |
-// +---------------------------------------+
-
-//Inserts a counter container as a child of the root atom.
-type InsertCounter struct{}
-
-func (v InsertCounter) AtomPriority() int { return insertCounterPriority }
-func (v InsertCounter) MarshalJSON() ([]byte, error) {
-	return json.Marshal("insert counter container")
-}
-
-func (v InsertCounter) String() string { return "Counter: " }
-
-func (v InsertCounter) ValidateChild(child atm.AtomValue) error {
-	switch child.(type) {
-	case InsertAdd, Delete:
-		return nil
-	default:
-		return fmt.Errorf("invalid atom value after InsertCounter: %T (%v)", child, child)
-	}
-}
-
-// InsertCounter inserts a Counter container after the root and advances the cursor.
-func (t *CausalTree) InsertCounter() error {
-	t.Cursor = atm.AtomID{}
-	atomID, err := t.addAtom(InsertCounter{})
-	t.Cursor = atomID
-	return err
-}
-
 // +------------+
 // | Conversion |
 // +------------+
-
-func toString(data interface{}) string {
-	switch v := data.(type) {
-	case int32:
-		return strconv.FormatInt(data.(int64), 10)
-	case string:
-		return data.(string)
-	case []interface{}:
-		var resultingStringBuilder strings.Builder
-		for _, element := range data.([]interface{}) {
-			resultingStringBuilder.WriteString(toString(element))
-		}
-		return resultingStringBuilder.String()
-	default:
-		panic(fmt.Sprintf("ToString: invalid json element (%T)", v))
-	}
-}
 
 // ToString returns a string representation of the CausalTree t.
 func (t *CausalTree) ToString() string {
@@ -888,7 +564,7 @@ func (t *CausalTree) ToString() string {
 		panic(fmt.Sprintf("ToString: %v", err))
 	}
 
-	return toString(data)
+	return conversion.ToString(data)
 
 }
 
@@ -934,27 +610,4 @@ func (t *CausalTree) ToJSON() ([]byte, error) {
 		panic(fmt.Sprintf("ToJSON: %v", err))
 	}
 	return finalJSON, nil
-}
-
-// +-----------+
-// | Utilities |
-// +-----------+
-
-// Provides a random MAC address.
-func randomMAC() []byte {
-	mac := make([]byte, 6)
-	if _, err := io.ReadFull(rand.Reader, mac); err != nil {
-		panic(err.Error())
-	}
-	return mac
-}
-
-// Create UUIDv1, using local timestamp as lower bits and random MAC.
-func randomUUIDv1() uuid.UUID {
-	uuid.SetNodeID(randomMAC())
-	id, err := uuid.NewUUID()
-	if err != nil {
-		panic(fmt.Sprintf("creating UUIDv1: %v", err))
-	}
-	return id
 }
