+++ src/causal_tree/operations_priority.go
@@ -1,0 +1,12 @@
+package causal_tree
+
+// +--------------------------------------+
+// + Operations - Atom Priority constants |
+// +--------------------------------------+
+const (
+	insertCharPriority    = 0
+	insertStrPriority     = 30
+	deletePriority        = 100
+	insertCounterPriority = 30
+	insertAddPriority     = 30
+)
